fix some bugs

+++ HousingService/Offers/Commands/AddOfferCommand.cs
@@ -12,7 +12,7 @@ namespace HousingService.Offers.Commands
         public OfferStatus Status { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public decimal PricePerDay { get; set; }
+        public double PricePerDay { get; set; }
         
         public sealed class AddOfferCommandHandler : IRequestHandler<AddOfferCommand, Unit>
         {
